requirements.txt
pandas>=2.0,<3.0
numpy>=1.24
pyyaml>=6.0
